+++ src/index.js
@@ -1,5 +1,5 @@
 export { readColumnIndex, readOffsetIndex } from './indexes.js'
-export { parquetMetadata, parquetMetadataAsync, parquetSchema } from './metadata.js'
+export { parquetMetadataAsync, parquetSchema } from './metadata.js'
 export { parquetRead, parquetReadObjects } from './read.js'
 export { parquetQuery } from './query.js'
 export { snappyUncompress } from './snappy.js'